implemented user configuration feature

+++ FilesCleaner.cs
@@ -1,4 +1,5 @@
-﻿using System.Timers;
+﻿using System.Reflection;
+using System.Timers;
 
 namespace RedundantFilesCleanerService
 {
@@ -8,18 +9,60 @@ namespace RedundantFilesCleanerService
         public FilesCleaner()
         {
             _timer = new System.Timers.Timer(1000) { AutoReset = true };
-            _timer.Elapsed += voidDeleteFiles;
+            _timer.Elapsed += DeleteFiles;
         }
-        private void voidDeleteFiles(object sender, ElapsedEventArgs e)
+        private void DeleteFiles(object sender, ElapsedEventArgs e)
         {
-
-            DirectoryInfo directoryInfo = new DirectoryInfo("C:\\WebReports\\Estimate Template");
-            var files = directoryInfo.EnumerateFiles();
-
-            foreach (var file in files)
+            var exPath = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
+            var path = exPath + "\\directoryConfig.txt";
+          
+           if(File.Exists(path))
             {
-                file.Delete();
+                string[] directoryPathList = File.ReadAllLines(path);
+                if(directoryPathList.Length > 0)
+                {
+                    var userCustomDaysStr = directoryPathList.Length > 1 ? directoryPathList[0] : string.Empty;
+                    directoryPathList = directoryPathList.Length > 1 ? directoryPathList.Skip(1).ToArray() : directoryPathList;                   
+                    
+                    foreach (string directoryPath in directoryPathList)
+                    {
+
+                        DirectoryInfo directoryInfo = new DirectoryInfo(directoryPath);
+                        if(directoryInfo.Exists)
+                        {
+                            var files = directoryInfo.EnumerateFiles();
+                            var directories = directoryInfo.EnumerateDirectories();
+                            var olderThanDays = -355;
+
+                            if (!string.IsNullOrWhiteSpace(userCustomDaysStr) && int.TryParse(userCustomDaysStr, out int userCustomDays))
+                            {
+                                olderThanDays = userCustomDays > 0 ? 0 - userCustomDays : userCustomDays;
+                            }
+                            if (olderThanDays < 0)
+                            {
+                                foreach (var file in files.Where(e => e.LastWriteTime < DateTime.Now.AddDays(olderThanDays) && e.CreationTime < DateTime.Now.AddDays(olderThanDays)))
+                                {
+                                    if (file.Exists)
+                                    {
+                                        file.Delete();
+                                    }
+                                }
+                                foreach (var directory in directories.Where(e => e.LastWriteTime < DateTime.Now.AddDays(olderThanDays) && e.CreationTime < DateTime.Now.AddDays(olderThanDays)))
+                                {
+                                    if (directory.Exists)
+                                    {
+                                        directory.Delete(true);
+                                    }
+                                }
+                            }
+                           
+                        }
+                      
+                    }
+                }
+               
             }
+         
         }
 
         public void Start()
